backend/requirements.txt
# FastAPI AI Chat Backend

## Core Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic-settings>=2.0.3
email-validator>=2.1.0
aiofiles>=24.1.0

## RAG and Vector Database
chromadb>=1.1.1
sentence-transformers>=5.1.1
more-itertools>=10.1.0

## Web Scraping
playwright>=1.45.0
html2text>=2024.2.26
requests>=2.31.0

## Document Parsing
PyPDF2>=3.0.1
pdfplumber>=0.10.3
python-docx>=1.1.0

## Google AI
google-genai>=0.1.0
